+++ src/components/AvatarGroup/AvatarGroupCounterTooltipContent.tsx
@@ -6,7 +6,6 @@ import AvatarGroupCounterTooltipContentVirtualizedList from "./AvatarGroupCounte
 import { avatarRenderer } from "./AvatarGroupCounterTooltipHelper";
 import VibeComponentProps from "../../types/VibeComponentProps";
 import { AvatarType } from "../Avatar/AvatarConstants";
-import { FlexDirection, FlexGap } from "../Flex/FlexConstants";
 import styles from "./AvatarGroupCounterTooltipContent.module.scss";
 
 interface AvatarGroupCounterTooltipContentProps extends VibeComponentProps {
@@ -67,9 +66,8 @@ const AvatarGroupCounterTooltipContent: React.FC<AvatarGroupCounterTooltipConten
     className: displayAsGrid
       ? cx(styles.scrollableContainer, styles.tooltipContainer, styles.tooltipGridContainer, className)
       : cx(styles.scrollableContainer, styles.tooltipContainer, className),
-    // TODO replace with enum when Flex is converted to TS
-    direction: displayAsGrid ? FlexDirection.ROW : FlexDirection.COLUMN,
-    gap: displayAsGrid ? FlexGap.XS : FlexGap.SMALL,
+    direction: displayAsGrid ? Flex.directions.ROW : Flex.directions.COLUMN,
+    gap: displayAsGrid ? Flex.gaps.XS : Flex.gaps.SMALL,
     wrap: displayAsGrid
   };