+++ src/components/Dropdown/Dropdown.jsx
@@ -3,7 +3,7 @@ import { ComponentDefaultTestId, getTestId } from "../../tests/test-ids-utils";
 import cx from "classnames";
 import { SIZES } from "../../constants/sizes";
 import React, { forwardRef, useCallback, useMemo, useRef, useState } from "react";
-import Select, { components } from "react-select";
+import Select, { components, createFilter } from "react-select";
 import AsyncSelect from "react-select/async";
 import { noop as NOOP } from "lodash-es";
 import { WindowedMenuList } from "react-windowed-select";
@@ -14,10 +14,16 @@ import OptionComponent from "./components/option/option";
 import SingleValueComponent from "./components/singleValue/singleValue";
 import ClearIndicatorComponent from "./components/ClearIndicator/ClearIndicator";
 import MultiValueContainer from "./components/MultiValueContainer/MultiValueContainer";
-import { ADD_AUTO_HEIGHT_COMPONENTS, defaultCustomStyles, DROPDOWN_ID } from "./DropdownConstants";
+import {
+  ADD_AUTO_HEIGHT_COMPONENTS,
+  defaultCustomStyles,
+  DROPDOWN_ID,
+  DROPDOWN_CHIP_COLORS,
+  DROPDOWN_MENU_PLACEMENT,
+  DROPDOWN_MENU_POSITION
+} from "./DropdownConstants";
 import generateBaseStyles, { customTheme } from "./Dropdown.styles";
 import Control from "./components/Control/Control";
-import { DROPDOWN_CHIP_COLORS } from "./dropdown-constants";
 import menuStyles from "./components/menu/menu.module.scss";
 import styles from "./Dropdown.module.scss";
 
@@ -81,6 +87,8 @@ const Dropdown = forwardRef(
       ariaLabel,
       tabSelectsValue = true,
       popupsContainerSelector,
+      filterOption,
+      menuPosition,
       "data-testid": dataTestId
     },
     ref
@@ -340,6 +348,7 @@ const Dropdown = forwardRef(
         maxMenuHeight={maxMenuHeight}
         menuPortalTarget={overrideMenuPortalTarget}
         menuPlacement={menuPlacement}
+        menuPosition={menuPosition}
         menuIsOpen={!readOnly && menuIsOpen}
         tabIndex={tabIndex}
         id={id}
@@ -352,6 +361,7 @@ const Dropdown = forwardRef(
         isLoading={isLoading}
         loadingMessage={loadingMessage}
         tabSelectsValue={tabSelectsValue}
+        filterOption={filterOption}
         {...asyncAdditions}
         {...additions}
       />
@@ -359,8 +369,15 @@ const Dropdown = forwardRef(
   }
 );
 
-Dropdown.size = SIZES;
-Dropdown.chipColors = DROPDOWN_CHIP_COLORS;
+Object.assign(Dropdown, {
+  // TODO Deprecate Dropdown.size in the next major version - use Dropdown.sizes instead
+  size: SIZES,
+  sizes: SIZES,
+  chipColors: DROPDOWN_CHIP_COLORS,
+  menuPlacements: DROPDOWN_MENU_PLACEMENT,
+  menuPositions: DROPDOWN_MENU_POSITION,
+  createFilter: createFilter
+});
 
 Dropdown.defaultProps = {
   className: "",
@@ -377,7 +394,8 @@ Dropdown.defaultProps = {
   onInputChange: NOOP,
   searchable: true,
   options: [],
-  menuPlacement: "bottom",
+  menuPlacement: Dropdown.menuPlacements.BOTTOM,
+  menuPosition: Dropdown.menuPositions.ABSOLUTE,
   noOptionsMessage: NOOP,
   clearable: true,
   size: SIZES.MEDIUM,
@@ -396,7 +414,8 @@ Dropdown.defaultProps = {
   readOnly: false,
   isLoading: false,
   loadingMessage: undefined,
-  ariaLabel: undefined
+  ariaLabel: undefined,
+  filterOption: undefined
 };
 
 Dropdown.propTypes = {
@@ -489,7 +508,12 @@ Dropdown.propTypes = {
   /**
    * Default placement of the Dropdown menu in relation to its control. Use "auto" to flip the menu when there isn't enough space below the control.
    */
-  menuPlacement: PropTypes.oneOf(["bottom", "top", "auto"]),
+  menuPlacement: PropTypes.oneOf(Object.values(DROPDOWN_MENU_PLACEMENT)),
+  /**
+   * The CSS position value of the menu, when "fixed" extra layout management might be required
+   * Fixed position can be used to solve the issue of positioning Dropdown inside overflow container like Modal or Dialog
+   */
+  menuPosition: PropTypes.oneOf(Object.values(DROPDOWN_MENU_POSITION)),
   /**
    * If set to true, the dropdown will be in Right to Left mode
    */
@@ -526,7 +550,7 @@ Dropdown.propTypes = {
     })
   ]),
   /**
-   * Select menu size from `Dropdown.size` - Dropdown.size.LARGE | Dropdown.size.MEDIUM | Dropdown.size.SMALL
+   * Select menu size from `Dropdown.size` - Dropdown.sizes.LARGE | Dropdown.sizes.MEDIUM | Dropdown.sizes.SMALL
    */
   size: PropTypes.string,
   /**
@@ -632,7 +656,12 @@ Dropdown.propTypes = {
   /**
    * Overrides the built-in logic of tab selecting value (default: true)
    */
-  tabSelectsValue: PropTypes.bool
+  tabSelectsValue: PropTypes.bool,
+  /**
+   * Overrides the build-in search filter logic - https://react-select.com/advanced#custom-filter-logic
+   * createFilter function is available at Dropdown.createFilter
+   */
+  filterOption: PropTypes.oneOfType([PropTypes.func, PropTypes.object])
 };
 
 export default Dropdown;