+++ src/components/TextField/TextField.tsx
@@ -198,6 +198,8 @@ const TextField: VibeComponent<TextFieldProps, unknown> & {
       }
     }, [inputRef, autoFocus]);
 
+    const isIconContainerClickable = onIconClick !== NOOP || clearOnIconClick;
+
     return (
       <div
         className={cx(styles.textField, wrapperClassName, {
@@ -260,7 +262,8 @@ const TextField: VibeComponent<TextFieldProps, unknown> & {
             <Clickable
               className={cx(styles.iconContainer, {
                 [styles.iconContainerHasIcon]: hasIcon,
-                [styles.iconContainerActive]: isPrimary
+                [styles.iconContainerActive]: isPrimary,
+                [styles.iconContainerClickable]: isIconContainerClickable
               })}
               onClick={onIconClickCallback}
               tabIndex={onIconClick !== NOOP && inputValue && iconName.length && isPrimary ? "0" : "-1"}
@@ -278,7 +281,8 @@ const TextField: VibeComponent<TextFieldProps, unknown> & {
             <Clickable
               className={cx(styles.iconContainer, {
                 [styles.iconContainerHasIcon]: hasIcon,
-                [styles.iconContainerActive]: isSecondary
+                [styles.iconContainerActive]: isSecondary,
+                [styles.iconContainerClickable]: isIconContainerClickable
               })}
               onClick={onIconClickCallback}
               tabIndex={!shouldFocusOnSecondaryIcon ? "-1" : "0"}