+++ src/components/ThemeProvider/__tests__/themeProvider-tests.jest.js
@@ -96,12 +96,6 @@ describe("ThemeProvider tests", () => {
             "primary-hover-color": "#ad4007",
             "primary-selected-color": "#6d2703",
             "primary-selected-hover-color": "#491b04"
-          },
-          hacker: {
-            "primary-color": "#d14903",
-            "primary-hover-color": "#ad4008",
-            "primary-selected-color": "#6d2704",
-            "primary-selected-hover-color": "#491b05"
           }
         }
       };
@@ -109,8 +103,7 @@ describe("ThemeProvider tests", () => {
       expect(css).toBe(
         `.light-app-theme .${ADDITIONAL_STRING_SELECTOR}.${theme.name} {--primary-color: #d14900;--primary-hover-color: #ad4005;--primary-selected-color: #f8dccf;--primary-selected-hover-color: #f1d3c4;}\n` +
           `.dark-app-theme .${ADDITIONAL_STRING_SELECTOR}.${theme.name} {--primary-color: #d14901;--primary-hover-color: #ad4006;--primary-selected-color: #6d2702;--primary-selected-hover-color: #491b03;}\n` +
-          `.black-app-theme .${ADDITIONAL_STRING_SELECTOR}.${theme.name} {--primary-color: #d14902;--primary-hover-color: #ad4007;--primary-selected-color: #6d2703;--primary-selected-hover-color: #491b04;}\n` +
-          `.hacker_theme-app-theme .${ADDITIONAL_STRING_SELECTOR}.${theme.name} {--primary-color: #d14903;--primary-hover-color: #ad4008;--primary-selected-color: #6d2704;--primary-selected-hover-color: #491b05;}\n`
+          `.black-app-theme .${ADDITIONAL_STRING_SELECTOR}.${theme.name} {--primary-color: #d14902;--primary-hover-color: #ad4007;--primary-selected-color: #6d2703;--primary-selected-hover-color: #491b04;}\n`
       );
     });